# cmd/glue/api/fpixel.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Flame Pixels Controller",
    "version": "0.1.0"
  },
  "host": "localhost:8080",
  "schemes": [
    "http"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/v1/sensor/{id}/raw": {
      "get": {
        "summary": "SensorRawData",
        "operationId": "GetSensorRawData",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/fpixelsSensorRawDataResponse"
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "tags": [
          "FlamePixels"
        ]
      }
    },
...
  },
  "definitions": {
...
    "fpixelsSensorRawDataResponse": {
      "type": "object",
      "properties": {
        "payload": {
          "type": "string"
        }
      }
    }
  }
}